Add a batch script to install the WSGI FastCGI handler.
Unfortunately, I can't get this to work in the `web.config` file yet
despite the docs indicating this should be possible.

If we can't get this to work in `web.config`, then we'll have to run
this command on install.  Unfortunately, the `IIS_BIN` environment
variable was only available when I had `runCommand` working so if I
can't get that back then we'll have to sniff whether IIS or IIS
Express is the place we should get `appcmd.exe` from.